fix poll
