+++ frontend/src/features/admin/index.ts
@@ -1,2 +1,3 @@
+import "./active-crawls-badge";
 import "./stats";
 import "./super-admin-banner";
