+++ frontend/src/index.ts
@@ -3,7 +3,6 @@ import "./global";
 
 import { provide } from "@lit/context";
 import { localized, msg, str } from "@lit/localize";
-import { Task } from "@lit/task";
 import type {
   SlDialog,
   SlDrawer,
@@ -15,7 +14,6 @@ import { ifDefined } from "lit/directives/if-defined.js";
 import { until } from "lit/directives/until.js";
 import { when } from "lit/directives/when.js";
 import isEqual from "lodash/fp/isEqual";
-import queryString from "query-string";
 
 import "./components";
 import "./features";
@@ -35,9 +33,7 @@ import AuthService, {
 import { BtrixElement } from "@/classes/BtrixElement";
 import type { NavigateEventDetail } from "@/controllers/navigate";
 import type { NotifyEventDetail } from "@/controllers/notify";
-import type { APIPaginatedList } from "@/types/api";
 import { type Auth } from "@/types/auth";
-import type { Crawl } from "@/types/crawler";
 import {
   translatedLocales,
   type TranslatedLocaleEnum,
@@ -69,8 +65,6 @@ export type APIUser = {
 export interface UserGuideEventMap {
   "btrix-user-guide-show": CustomEvent<{ path?: string }>;
 }
-
-const POLL_INTERVAL_SECONDS = 30;
 
 @customElement("browsertrix-app")
 @localized()
@@ -113,24 +107,6 @@ export class App extends BtrixElement {
 
   @query("#userGuideDrawer")
   private readonly userGuideDrawer!: SlDrawer;
-
-  private readonly activeCrawlsTotalTask = new Task(this, {
-    task: async () => {
-      return await this.getActiveCrawlsTotal();
-    },
-    args: () => [] as const,
-  });
-
-  private readonly pollTask = new Task(this, {
-    task: async ([crawls]) => {
-      if (!crawls) return;
-
-      return window.setTimeout(() => {
-        void this.activeCrawlsTotalTask.run();
-      }, POLL_INTERVAL_SECONDS * 1000);
-    },
-    args: () => [this.activeCrawlsTotalTask.value] as const,
-  });
 
   get orgSlugInPath() {
     return this.viewState.params.slug || "";
@@ -186,12 +162,6 @@ export class App extends BtrixElement {
     });
 
     this.startSyncBrowserTabs();
-  }
-
-  disconnectedCallback(): void {
-    super.disconnectedCallback();
-
-    window.clearTimeout(this.pollTask.value);
   }
 
   private attachUserGuideListeners() {
@@ -629,14 +599,7 @@ export class App extends BtrixElement {
                     @click=${this.navigate.link}
                   >
                     ${msg("Active Crawls")}
-                    ${when(
-                      this.activeCrawlsTotalTask.value,
-                      (total) => html`
-                        <btrix-badge variant=${total > 0 ? "primary" : "blue"}>
-                          ${this.localize.number(total)}
-                        </btrix-badge>
-                      `,
-                    )}
+                    <btrix-active-crawls-badge></btrix-active-crawls-badge>
                   </a>
                 </div>
               `
@@ -1182,17 +1145,5 @@ export class App extends BtrixElement {
 
   private clearSelectedOrg() {
     AppStateService.updateOrgSlug(null);
-  }
-
-  private async getActiveCrawlsTotal() {
-    const query = queryString.stringify({
-      pageSize: 1,
-    });
-
-    const data = await this.api.fetch<APIPaginatedList<Crawl>>(
-      `/orgs/all/crawls?${query}`,
-    );
-
-    return data.total;
   }
 }
